service-2-grpc-server/requirements.txt
grpcio-tools
pandas
flask
tabulate